Pass instances of DependencyMirrors instead of SwiftPMConfig to APIs
Move SwiftPMConfig type to Workspace module

Move DependencyMirrors type to PackageGraph module
